Fix #248. This was quite a rabbit. Here is what is contained:
There were some undocumented solve! and backsolve! which did solving
and (ff)lu factorization at the same time. I removed them and replaced
them by solve_fflu, solve_fflu_precomp, solve_lu, solve_lu_precomp.
This has two advantages:
- It does the same as flint.
- One can solve multiple times with the same left side.
- I can get the determinant.

Also the solve_interpolation used interpolated solutions, where the
interpolated determinant was zero. This is not quite correct if
solve_with_det is allowed any solution together with the determinant.

The generic solve_with_det called solve_rational, which is confusing
since there is no reason solve_rational has to return the determinant
as the denominator (it is only a coincidence that it is the same for
fmpz_mat).

The timings of the benchmarks are not effected.