tests: use an easy to include setup file
When wanting to run a specific test/testset, it's easy to copy-pase it
into the REPL, but it can be annoying to find the setup-code, like
constants or functions defined usually at the top of the test file.
All this setup code is here put in test/setup.jl, so that it's always
predictable (and easy) what one has to do to inject to required
dependencies in the REPL: `include("test/setup.jl")`.