interpret rand(matspace, 2) as creating a Vector
Currently, rand(matspace, 2) is ambiguous; it could mean:
* rand(make(matspace), 2), which creates a vector
* rand(make(matspace, 2)) == rand(make(matspace, make(base_ring(matspace), 2))),
  i.e. the "2" is used as a make-argument to specify how to generate random
  elements of the matrix

We choose the first interpretation here, as we are used to interpret
a call like `rand(X, 2)` as creating a vector, `rand(X, 2, 3)` as
creating a matrix and so on. On the other hand, there are almost no
(or not at all) cases in Oscar where an integer is used as
an argument to make.