make Integers, Rationals and Floats immutable structs
Instances of these types are meant to be singleton objects
via the caching mechanism, but as they no fields, the
same can be achieved by making them immutable
(having no fields implies that they are bitstypes and
automatically have the singleton property, like e.g. `nothing`;
this also means that comparison should be at leas as cheap
as comparing pointers like was previously the case when mutable).

Moreover, in a `Distributed` setting, uniqueness of instances
was broken, in particular instances from two process would
compare unequal (an alternative fix for that would be to
define `==`).