Remove adjoint method calling transpose
In Julia, adjoint is defined to be conjugate transposition. There are rings
(e.g. finite fields of even degree over their base fields) that admit a
canonical involution and where the notion of conjugate transposition is
used as well, and standard in the literature.

As such, I think it is misleading and potentially dangerous to simply map
`adjoint` to `transpose`.

One drawback of removing it is that the notation `a'` won't work anymore for
transposing a matrix.