Add @attr to cache function return values
Using our attributes system, it is easy to cache the return value
of expensive functions: Just turn

    function myattr(obj::Foo)
        # ... expensive computation
        return result::fmpz
    end

into something like this (provided that `Foo` supports attribute storing)

    function myattr(obj::Foo)
      return get_attribute!(obj, :myattr) do
        # ... expensive computation
        return result::fmpz
      end
    end

However, doing this repeatedly quickly gets tedious. It also comes with a
catch: it prevents type inference on the return value of `myattr`: while in
the first example `Base.return_types(myattr)` produces `[fmpz]`, in the second
example it yields `[Any]`. This can be resolved by manually adding a suitable type
assertion on the return value of `myattr`, but that's easy to forget.

The `@attr` macro resolves this and more. With it, you can write the following
code to get the same caching as in the second example, but with a correctly
inferred return type:

    @attr function myattr(obj::Foo)
        # ... expensive computation
        return result::fmpz
    end

Under the hood, this works by renaming the function doing the actual
computation, then calling `Base.return_types` to determine its return type(s).
Then a wrapper function under the original name is created which invokes
`get_attribute!` suitably and applies a type assertion based on the return
type(s) that were just computed, i.e., basically what was sketched above, but
all done automatically.