Revise @attr return type handling
The old approach using `Base.return_types` was fatally flawed:
we can't always correctly deduce the return type of code at the
time it is defined, as e.g. it might reference methods that are
not even defined yet.

So we remove this code. Instead, an optional first argument is
added to the `@attr` macro which now is used to specify a
type assertion on the type of the attribute.