Remove `get_field` and `set_field!`
There were meant to provide additional generality to maps in Oscar, and
allow writing "more generic" code. In practice, as far as I can tell no
public code outside of AbstractAlgebra ever used them. As to writing
generic code: the idea, as I understand it (which may be wrong) seems to
be that certain generic map implementation can use it to forward
"arbitrary" map functionality, even beyond what they were made for. But
the same could be done by adding explicit forwarding methods, as indeed
this patch does for `domain` and `codomain`. This also means that not
all map types must actually have fields `domain` and `codomain`, as was
the case with the `get_field` mechanism.