Deflate and inflate exponents of only some variables
Given an MPolyElem f and a list of some of the variables in f,
deflate(f, vars, shift, defl) returns a polynomial obtained by deflating
only the specified variables. vars can be an array of either generators
or variable indices. shift and defl var arrays of integers, one for each
of the specified variables. inflate(f, vars, shift, defl) is defined
similarly.