Improve type stability in `rref!`
... and thereby also resulting method invalidations, by adding
type assertion guaranteeing that nrows and ncols return an `Int`.

In an ideal world Julia could always deduce this from type inference.
But in practice, this may sometimes be impossible, namely if the
type of the matrix `A` is not fully known (e.g. because the code
invoking `rref!`, does not know it -- in other words, if somewhere
above in the call stack the type is not known).