Add abstract `(::MatSpace)(::MatrixElem)` method (#1268)
* Add abstract `(::MatSpace)(::MatrixElem)` method

This makes the method for `Generic.MatSpace` superfluous and also those
of other `MatSpace` implementations. Especially the stub implementations
in `Nemo.jl`.

* Fix new tests

For coercing a matrix into a matrix with other coefficient types, I
prefer not to test which `Exception` it throws. Unless a specialized
method exists this currently throws a `MethodError`. For the
`DomainError` I provide an example where the error is definitely
mathematically necessary.

* Apply renaming to tests