Fix PrettyPrinting corner case, add more tests
Add some examples from the OSCAR dev docs to make sure they match
reality.

Also fix a bug where in some situations the pretty printing feature
caused a trailing 0 to be printed (e.g. when ending printing with
a `Dedent()`), as in this example:

    julia> show(stdout, MIME"text/plain"(), F)
    Number field with defining polynomial x^2 + 3
      over rational field0

I tried to add a test case for this, but the issue only happens
with `stdout`, not with an `IOBuffer`