Modify testset "Generic.Ideal.addition"
Don't use random examples in the multivariate case, as some of these can
lead to *really* horribly bad runtimes. Instead use a fixed list of
examples.

For the univariate cases and the ideals in Z, I did not change anything
so far, as those should not suffer from the same problems (famous last
words...)