Restrict VarNames union type
We don't `VarName` as part of `VarNAmes`. This removes a few call variations
of `polynomial_ring` etc., that we did not actually agree to put in (they were
likely a leftover of an earlier iteration of the recent varname changes).

This allows us to remove some `polynomial_ring` methods that were added purely
for disambiguation, but which caused a type instability in `polynomial_ring`
in Julia 1.6 (but not in 1.9).