Don't swap rows and columns in `kernel(::FPModuleHomomorphism)`
I don't know what 'upper triangular wrt original data' is supposed to
mean.
The function `AbstractAlgebra.kernel` does not care whether it
gets an upper triangular matrix, so I don't see a point in trying to
maintain the structure (and I don't see how the original code does this
by reversing the rows of `M`).
The code that constructs the generators for the submodule appeared to
assume something on the structure of `AbstractAlgebra.kernel` where we
decidedly do not guarantee anything (anymore?).