move code from `Oscar.jl/experimental/GModule/Misc.jl` here
The idea is to add this code to AbstractAlgebra.jl,
and then to remove it from Oscar.jl.

Once we agree what shall be really added here,
tests and documentation has to be added.
(I am quite sure that I have put certain parts into the wrong files,
and that some functions have to be changed.)