Fix and document AbstractAlgebra.ispower_moduli
The prior code did not do any wrong computations, it just
neglected to actually use the third set of moduli/residues
tables for the prime 7 due to a typo, which this patch fixes.

This patch also cleans up the handling of even 'a' and adds
a bunch of comments to explain what is going on.

Finally, it also extends the check mod 8 to even n. The prior
check here was a bit hackish in that it hardcoded the assumption
that moduli5[1] is 8.