Prefer symbols for VarNames over strings
Both are valid and allowed and will remain so. But using symbols overall
avoids a bunch of allocations, so it makes sense to adjust our internal
code as well as user-facing documentation to use symbols to set a good
precedent.

To demonstrate the effect, here a benchmark with strings:

    julia> @btime polynomial_ring(QQ, ["x", "y"]);
      756.274 ns (27 allocations: 1.22 KiB)

Here the same with symbols:

    julia> @btime polynomial_ring(QQ, [:x, :y]);
      535.963 ns (20 allocations: 896 bytes)

Of course code where this noticeable would be code that calls polynomial_ring
far too many times, and should be changed to not do *that*.