Add is_euclidean_type trait
The idea is that this true for ring / ring elem types that
explicitly support a euclidean division with remainder, and
ideally also a euclidean degree (to be defined in the interface).
This is just a draft to have a point for discussion and
experiments: it is missing many things, including documentation as
to what is_euclidean_type(x)==true means, which promises this
incurrs; and of course also code needs to be adapted to use it
(e.g. the residue ring or ideal constructions or whatnot that only
work sensible for euclidean rings should check this.