Support function call syntax for evaluate for more rings (#2090)
Now rational function fields, fraction fields and factored
fraction fields allow using "function call syntax" as a shorthand
for `evaluate`.

This also resolves an issue in the RationalFunctionField tests
for `evaluate` which did not actually work as intended, as the
elements being evaluated action ended up being of type `FracElem`.