Fix in(key_value::Pair, io::IOCustom)
The three argument method it used to delegate to does not exist. Most
likely it was a copy & pasted & modified version of this code from the
Julia stdlib:

    in(key_value::Pair, io::IOContext) = in(key_value, io.dict, ===)

The three arg in method does indeed exist for Dict types, but not
for IO types. The new code clearly does the intended thing.