Resolve a bunch of JET warnings (#2143)
* Fix is_finite(::MatRing)

* Refactor FinFieldIterator to improve type stability

With this change, Julia can prove that `iterate(GF(p))` will never
return `nothing` (finite fields are not empty). This resolves a JET
warning.

* Make underscorify return type stable

It always returns a String, but before the Julia compiler could not
prove that and thought it might sometimes return a SubString.

* Fix in(key_value::Pair, io::IOCustom)

The three argument method it used to delegate to does not exist. Most
likely it was a copy & pasted & modified version of this code from the
Julia stdlib:

    in(key_value::Pair, io::IOContext) = in(key_value, io.dict, ===)

The three arg in method does indeed exist for Dict types, but not
for IO types. The new code clearly does the intended thing.

* Make polynomial_to_power_sums type stable

... to resolve yet another JET warning

* Fix broken references to Generic.FunctionalMap

An alternate fix would have been to just replace `FunctionalMap` by
`Generic.FunctionalMap` in that code; but the code really belongs into
src/generic/Map.jl anyway, so let's move it there.

* Fix and speed up content(::SparsePoly)

* fixup! Fix broken references to Generic.FunctionalMap

* Fix type instability in is_univariate_with_data

Reported by JET

* Fix JET warning in Perm constructor

A range 1:n was constructed with n of some type T <: Integer, but 1
an Int. The code used `collect(T, 1:n)` to deal with this inconsistency,
but the proper fix is to just create a range of the correct type.

* Fix JET warning in normalise_keyword_arguments

The code was appending a generator expression, that lead to some JET
warnings deep in the Julia stdlib. In any case it seems fine to just
use a list comprehension here instead.

* Tweak Fac constructor to ensure Julia knows f.fac is always set