Fix JET complaint in sqrt_classical
JET is concerned `g2` might be used without initialization.

However, in case of the method taking a `AbsPowerSeriesRingElem`
that access is in a loop over `1:prec - aval2 - 1` which is empty
unless `prec - aval2 - 1 >= 1`, i.e., `prec > aval2 + 1`.

Therefore `prec > aval2` is a necessary precondition, and so we
may move that loop inside the `if prec > aval2` block; or equivalently,
skip the code if `prec <= aval2`.

It is similar in the case of the `LaurentSeriesElem` and
`RelPowerSeriesRingElem` methods.