Expand the common code for ideal implementations (#2108)
This includes the addition of a `DefaultIdealSet` struct type
which can be used with zero overhead, and hopefully can simply
be used as parent type for all ideal implementations (in Hecke,
Oscar or wherever) down the road.

Also semi-documented the interfaces to be implemented (the list
is surely incomplete but better than nothing).